Start creating counterfactual menus.
