+++ src/webserver/web-gui/src/agents/Agent.css
@@ -1,7 +1,7 @@
 .Agent {
     display:flex;
     justify-content: space-between;
-    flex-direction: column;
+    flex-direction: row;
     padding-top: 2em;
     padding-left: 2em;
     padding-right: 2em;
@@ -25,6 +25,58 @@
     padding-left: 2em;
     padding-right: 2em;
     padding-bottom: 2em;
+}
+
+.AgentMain {
+    display:flex;
+    justify-content: space-between;
+    flex-direction: column;
+    padding-left: 2em;
+    padding-right: 2em;
+    padding-bottom: 2em;
+}
+
+.AgentHeader {
+    margin-top: -1.5em;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.AgentHeader h2 {
+    margin-top: 1.5em;
+    margin-bottom: 1em;
+    margin-right: 0.1em;
+}
+
+.AgentHeader button {
+    margin-left: auto;
+}
+
+.ModelChoice {
+    margin-left: 3em;
+    margin-top: 0.5em;
+    display: flex;
+    align-items: center;
+}
+
+.ModelChoice h4 {
+    margin-top: 3.5em;
+    margin-bottom: 3em;
+    margin-right: 1em;
+}
+
+.card {
+    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
+    transition: 0.3s;
+}
+
+.card:hover {
+    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
+}
+
+.card {
+    padding: 16px 16px 16px 16px;
 }
 
 .IntersectionArrival {
@@ -53,7 +105,7 @@
 .InitialDirection h4 {
     margin-top: 1.2em;
     margin-bottom: 1em;
-    margin-right: 0.3em;
+    margin-right: 1em;
 }
 
 .InitialDirection button {
