+++ src/webserver/web-gui/src/agents/Agents.css
@@ -10,37 +10,17 @@
     overflow:hidden;
 }
 
-.AgentHeader {
+.AgentsHeader {
     margin-top: -1.5em;
     display: flex;
     align-items: center;
 }
 
-.AgentHeader h2 {
+.AgentsHeader h2 {
     margin-top: 1.1em;
     margin-right: 0.1em;
 }
 
-.AgentHeader h3 {
-    margin-top: 1.5em;
-    margin-bottom: 1em;
-    margin-right: 0.1em;
-}
-
-.AgentHeader button {
+.AgentsHeader button {
     margin-left: auto;
-}
-
-.card {
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-    transition: 0.3s;
-}
-
-.card:hover {
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-}
-
-.card {
-    padding: 16px 16px 16px 16px;
-    margin-right: 30%;
 }
