+++ src/webserver/web-gui/src/agents/ModelChoice.js
@@ -41,11 +41,20 @@ class ModelChoice extends React.Component {
     ];
 
     return( 
-      <div>
-        <p> <br />Agent 0's Decision Model: </p>
+      <div className="ModelChoice">
+        <h4> Decision Model: </h4>
         <Select
           options={options} // Options to display in the dropdown
           defaultValue={{value: 'good_agent', label: 'Good Agent (Hard-Coded Behavior)'}} // Options to display in the dropdown
+          theme={theme => ({
+            ...theme,
+            borderRadius: 0,
+            colors: {
+              ...theme.colors,
+              primary25: '#DDCC77',
+              primary: '#999933',
+            },
+          })}
           onChange= {new_option => {this.handleChoice(new_option.value)}}
           />
       </div>);
